# proyecto-final/backend/InstaFilms/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: InstaFilms
# during dev, should point to your local machine 
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /movie:
    x-swagger-router-controller: movie.controller

    post:
      description: Add movie to db
      # used as the method name of the controller
      operationId: addMovie
      parameters:
        - in: body
          name: movie
          description: add movie
          schema:
            $ref: "#/definitions/Movie"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/GetMovieResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"  
  /movie/{id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: movie.controller
  
    get:
      description: Returns individual movie by id
      # used as the method name of the controller
      operationId: getMovieById
      parameters:
        - name: id
          in: path
          description: Movie to be returned
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GetMovieResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
    put:
      description: Updates movie by securityDefinitions
      operationId: updateMovieById
      parameters:
        - name: id 
          in: path
          description: Movie id to update
          required: true
          type: string
        - in: body
          name: movie
          description: The movie to update.
          schema: 
            $ref: "#/definitions/Movie"
      responses:
        "200": 
          description: Success
          schema:
            $ref: "#/definitions/GetMovieResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /movies/landing/:
    x-swagger-router-controller: movie.controller
    get:
    #Filtered by genre
      description: get the movie poster
      operationId: getMovieLanding
      parameters:
        - name: genre
          in: query
          description: optinal genres
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetMoviePosterResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /movies/title/:
    x-swagger-router-controller: movie.controller
    get:
      description: Returns titles ( LIKE +"OCEANS"+)
      # used as the method name of the controller
      operationId: getMovieTitle
      parameters:
        - name: title
          in: query
          description: find all movies containing that string in title
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GetMovieTitleResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  Movie:
    type: object
    properties:
      imdbid:
        type: string
        description: ID imdb.
      title:
        type: string
        description: Movie title.
      year:
        type: string
        format: date
        description: year movie was released.
      runtime:
        type: string
        description: Runetime in minutes.
      genre: 
        type: string
        description: Movie genre.
      director:
        type: string
        description: Movie director name.
      writer:
        type: string
        description: Movie writer.
      actors:
        type: string
        description: Main actors present in movie.
      plot:
        type: string
        description: Movie main sequence of events.
      poster:
        type: string
        description: Movie poster.
      status:
        type: boolean
        description: Is the movie available or not?

      

#Users
...
  GetMovieResponse:
      type: object
      properties:
        imdbid:
          type: string
          description: ID imdb.
        title:
          type: string
          description: Movie title.
        year:
          type: string
          format: date
          description: year movie was released.
        runtime:
          type: string
          description: Runetime in minutes.
        genre: 
          type: string
          description: Movie genre.
        director:
          type: string
          description: Movie director name.
        writer:
          type: string
          description: Movie writer.
        actors:
          type: string
          description: Main actors present in movie.
        plot:
          type: string
          description: Movie main sequence of events.
        poster:
          type: string
          description: Movie poster.
        status:
          type: boolean
          description: Is the movie available or not?
      required:
        - imdbid
        - title
        - year
        - runtime
        - genre
        - director
        - writer
        - actors
        - plot
        - poster
        - status
  GetMovieTitleResponse:
      type: object
      properties:
        title:
          type: string
          description: Movie title.
      required:
        - title
  GetMoviePosterResponse:
      type: object
      properties:
        IDimdb:
          type: string
          description: ID imdb.
        poster:
          type: string
          description: Movie poster.
      required:
        - IDimdb
        - poster
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...